Move handles to separate header
